requirements-test.txt
# Testing dependencies for Pentair IntelliCenter integration
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
ruff>=0.8.0
mypy>=1.13.0
